added Panel component
