+++ src/App.js
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+import Panel from "./Panel";
+
+class App extends Component {
+  render() {
+    return (
+      <div>
+        <Panel />
+      </div>
+    );
+  }
+}
+
+export default App;
